fix : 트로피 잘리는 문제 해결
