+++ src/1_Page/Team/ui/AwardList/index.tsx
@@ -2,57 +2,24 @@ import HoverTrophy from "./ui/HoverTrophy";
 import useGetTeamAwards from "../../../../3_Entity/Team/useGetTeamAwards";
 import React from "react";
 import useInfiniteScrollPaging from "../../../../4_Shared/model/useInfiniteScrollPaging";
+import useIndexAutoAnimate from "./model/useIndexAutoAnimate";
 
 const trophiesPerGroup = 5;
 
 const AwardList = () => {
-  const [currentIndex, setCurrentIndex] = React.useState<number>(0);
-
-  const animationTimerRef = React.useRef<ReturnType<typeof setTimeout> | null>(
-    null
-  );
-
   const [page, setPage] = React.useState<number>(1);
   const [teamAwards, hasMoreContent, loading] = useGetTeamAwards(page);
   const totalGroups = Math.ceil(teamAwards.length / trophiesPerGroup);
+
+  const [currentIndex, handleClickCurrentIndex] = useIndexAutoAnimate(
+    teamAwards.length
+  );
 
   const [observeRef] = useInfiniteScrollPaging(
     setPage,
     loading,
     hasMoreContent
   );
-
-  // 자동 애니메이션 설정
-  React.useEffect(() => {
-    // 애니메이션 타이머 설정
-    const startAnimation = () => {
-      if (animationTimerRef.current) {
-        clearInterval(animationTimerRef.current);
-      }
-
-      animationTimerRef.current = setInterval(() => {
-        setCurrentIndex((prev) => {
-          const next = prev + 1;
-
-          // 마지막 트로피에 도달하면 다음 페이지로 이동
-          if (next + 2 >= teamAwards.length) {
-            return 0;
-          }
-
-          return next;
-        });
-      }, 3000); // 3초마다 이동
-    };
-
-    startAnimation();
-
-    // 컴포넌트 언마운트 시 타이머 정리
-    return () => {
-      if (animationTimerRef.current) {
-        clearInterval(animationTimerRef.current);
-      }
-    };
-  }, [teamAwards.length]);
 
   return (
     <div className="flex flex-col justify-center items-center">
@@ -91,7 +58,7 @@ const AwardList = () => {
                 ? "bg-blue-500"
                 : "bg-gray-300"
             }`}
-            onClick={() => setCurrentIndex(index * trophiesPerGroup)}
+            onClick={() => handleClickCurrentIndex(index * trophiesPerGroup)}
           />
         ))}
       </div>
