+++ src/1_Page/Team/ui/AwardList/ui/HoverTrophy/index.tsx
@@ -15,12 +15,19 @@ const Trophy = ({
   const handleMouseEnter = (event: React.MouseEvent) => {
     setIsHovered(true);
     const rect = event.currentTarget.getBoundingClientRect();
+
+    const screenCenter = window.innerWidth / 2; // 화면 중앙
+    const offset = 60; // 추가 여백
+
     setHoverPosition({
-      x: rect.left + window.scrollX + rect.width / 2,
-      y: rect.top + window.scrollY + rect.height / 2 + 10, // 약간 아래로 조정
+      x:
+        rect.left +
+        window.scrollX +
+        rect.width / 2 +
+        (rect.left < screenCenter ? offset : -offset), // 좌우 방향 조정
+      y: rect.top + window.scrollY + rect.height, // 약간 아래로 조정
     });
   };
-
   const handleMouseLeave = () => {
     setIsHovered(false);
   };
